GODRIVER-2570 move Command to the component.go file
